requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
discord==1.0.1
discord.py==1.7.2
Flask==2.0.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
multidict==5.1.0
requests==2.25.1
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.0
yarl==1.6.3

Pillow~=8.2.0
gifmaker~=1.5